+++ R/netmeta-package.R
@@ -62,6 +62,7 @@
 #' 
 #' Additional methods and functions:
 #' \itemize{
+#' \item subgroup network meta-analysis (\code{\link{subgroup.netmeta}});
 #' \item information on network connectivity
 #'   (\code{\link{netconnection}});
 #' \item contribution of direct comparisons to network estimates