+++ R/print.netcomb.R
@@ -583,6 +583,8 @@ print.netcomb <- function(x,
     ##
     ## (d) Heterogeneity / inconsistency
     ##
+    print.I2 <- print.I2 & !is.na(I2)
+    #
     if (overall.hetstat) {
       if (print.tau2 | print.tau | print.I2 | print.Q)
         cat("\n")
@@ -613,8 +615,7 @@ print.netcomb <- function(x,
           text.hetstat <- paste0(
             text.hetstat,
             if (print.tau2 | print.tau) "; ",
-            if (!is.na(I2))
-              paste0(text.I2, " = ", round(I2, digits.I2), "%"),
+            paste0(text.I2, " = ", round(I2, digits.I2), "%"),
             if (print.I2.ci & (!(is.na(lower.I2) | is.na(upper.I2))))
               pasteCI(lower.I2, upper.I2,
                       digits.I2, big.mark, unit = "%"))